+++ benches/transform.rs
@@ -1,130 +1,132 @@
+#![allow(deprecated)]
+
 #[path = "support/macros.rs"]
 #[macro_use]
 mod macros;
 mod support;
 
 use criterion::{criterion_group, criterion_main, Criterion};
-use glam::{Isometry3A, Transform3A};
+use glam::{TransformRT, TransformSRT};
 use std::ops::Mul;
 use support::*;
 
-fn random_transform3a(rng: &mut PCG32) -> Transform3A {
-    Transform3A::from_scale_rotation_translation(
+fn random_transformsrt(rng: &mut PCG32) -> TransformSRT {
+    TransformSRT::from_scale_rotation_translation(
         random_nonzero_vec3(rng),
         random_quat(rng),
         random_vec3(rng),
     )
 }
 
-fn random_isometry3a(rng: &mut PCG32) -> Isometry3A {
-    Isometry3A::from_rotation_translation(random_quat(rng), random_vec3(rng))
+fn random_transformrt(rng: &mut PCG32) -> TransformRT {
+    TransformRT::from_rotation_translation(random_quat(rng), random_vec3(rng))
 }
 
 bench_unop!(
-    isometry3a_inverse,
-    "isometry3a inverse",
+    transformrt_inverse,
+    "TransformRT inverse",
     op => inverse,
-    from => random_isometry3a
+    from => random_transformrt
 );
 
 bench_unop!(
-    transform3a_inverse,
-    "transform3a inverse",
+    transformsrt_inverse,
+    "TransformSRT inverse",
     op => inverse,
-    from => random_transform3a
+    from => random_transformsrt
 );
 
 bench_binop!(
-    isometry3a_transform_point3,
-    "isometry3a transform point3",
+    transformrt_transform_point3,
+    "TransformRT transform point3",
     op => transform_point3,
-    from1 => random_isometry3a,
+    from1 => random_transformrt,
     from2 => random_vec3
 );
 
 bench_binop!(
-    isometry3a_transform_point3a,
-    "isometry3a transform point3a",
+    transformrt_transform_point3a,
+    "TransformRT transform point3a",
     op => transform_point3a,
-    from1 => random_isometry3a,
+    from1 => random_transformrt,
     from2 => random_vec3a
 );
 
 bench_binop!(
-    isometry3a_transform_vector3,
-    "isometry3a transform vector3",
+    transformrt_transform_vector3,
+    "TransformRT transform vector3",
     op => transform_vector3,
-    from1 => random_isometry3a,
+    from1 => random_transformrt,
     from2 => random_vec3
 );
 
 bench_binop!(
-    isometry3a_transform_vector3a,
-    "isometry3a transform vector3a",
+    transformrt_transform_vector3a,
+    "TransformRT transform vector3a",
     op => transform_vector3a,
-    from1 => random_isometry3a,
+    from1 => random_transformrt,
     from2 => random_vec3a
 );
 
 bench_binop!(
-    transform3a_transform_point3,
-    "transform3a transform point3",
+    transformsrt_transform_point3,
+    "TransformSRT transform point3",
     op => transform_point3,
-    from1 => random_transform3a,
+    from1 => random_transformsrt,
     from2 => random_vec3
 );
 
 bench_binop!(
-    transform3a_transform_point3a,
-    "transform3a transform point3a",
+    transformsrt_transform_point3a,
+    "TransformSRT transform point3a",
     op => transform_point3a,
-    from1 => random_transform3a,
+    from1 => random_transformsrt,
     from2 => random_vec3a
 );
 
 bench_binop!(
-    transform3a_transform_vector3,
-    "transform3a transform vector3",
+    transformsrt_transform_vector3,
+    "TransformSRT transform vector3",
     op => transform_vector3,
-    from1 => random_transform3a,
+    from1 => random_transformsrt,
     from2 => random_vec3
 );
 
 bench_binop!(
-    transform3a_transform_vector3a,
-    "transform3a transform vector3a",
+    transformsrt_transform_vector3a,
+    "TransformSRT transform vector3a",
     op => transform_vector3a,
-    from1 => random_transform3a,
+    from1 => random_transformsrt,
     from2 => random_vec3a
 );
 bench_binop!(
-    transform3a_mul_transform3a,
-    "transform3a mul transform3a",
+    transformsrt_mul_transformsrt,
+    "TransformSRT mul TransformSRT",
     op => mul,
-    from => random_transform3a
+    from => random_transformsrt
 );
 
 bench_binop!(
-    isometry3a_mul_isometry3a,
-    "isometry3a mul isometry3a",
+    transformrt_mul_transformrt,
+    "TransformRT mul TransformRT",
     op => mul,
-    from => random_isometry3a
+    from => random_transformrt
 );
 
 criterion_group!(
     benches,
-    isometry3a_inverse,
-    isometry3a_mul_isometry3a,
-    isometry3a_transform_point3,
-    isometry3a_transform_point3a,
-    isometry3a_transform_vector3,
-    isometry3a_transform_vector3a,
-    transform3a_inverse,
-    transform3a_mul_transform3a,
-    transform3a_transform_point3,
-    transform3a_transform_point3a,
-    transform3a_transform_vector3,
-    transform3a_transform_vector3a,
+    transformrt_inverse,
+    transformrt_mul_transformrt,
+    transformrt_transform_point3,
+    transformrt_transform_point3a,
+    transformrt_transform_vector3,
+    transformrt_transform_vector3a,
+    transformsrt_inverse,
+    transformsrt_mul_transformsrt,
+    transformsrt_transform_point3,
+    transformsrt_transform_point3a,
+    transformsrt_transform_vector3,
+    transformsrt_transform_vector3a,
 );
 
 criterion_main!(benches);